Refactoring and sort by due date
