+++ src/main/java/com/novare/sdaip/entity/Task.java
@@ -5,31 +5,34 @@ import java.util.Objects;
 
 /**
  * Task is a class which represents tasks to be added by users.
- * Task will be associated with a {@link Project}. It will have a title, status and a due date.
+ * It will have a title,a due date, status and a project .
  *
  * @author Neetu Jagya
+ * @version 1.0
  */
 public class Task {
 
     private String title;
     private Date dueDate;
     private TaskStatus status;
+    private String project;
 
-    private int taskId;
-    /**
-     * It is a default constructor.
-     */
-    public Task() {
+    @Override
+    public String toString() {
+        return "Task{" +
+                "title='" + title + '\'' +
+                ", dueDate=" + dueDate +
+                ", status=" + status +
+                ", project='" + project + '\'' +
+                '}';
     }
 
-    public Task(String title, Date dueDate, TaskStatus status, int taskId) {
+    public Task(String title, Date dueDate, TaskStatus status, String project) {
         this.title = title;
         this.dueDate = dueDate;
         this.status = status;
-        this.taskId = taskId;
+        this.project = project;
     }
-
-
 
     public String getTitle() {
         return title;
@@ -43,8 +46,5 @@ public class Task {
         return status;
     }
 
-    public int getTaskId() {
-        return taskId;
-    }
-
+    public String getProject() { return  project; }
 }
